+++ src/framework/compositors/CLCompositorAccumulate.ipp
@@ -1,7 +1,7 @@
 #ifndef CLCOMPOSITOR_ACCUMULATE_IPP
 #define CLCOMPOSITOR_ACCUMULATE_IPP
 
-template class CLCompositorAccumulate< uint > ;
+template class CLCompositorAccumulate< float > ;
 
 #endif // CLCOMPOSITOR_ACCUMULATE_IPP
